+++ lib/pokemon/view/pokemon.dart
@@ -1,9 +1,10 @@
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:pokemon_tracker/pokemon/bloc/pokemon_bloc.dart';
+import 'package:pokemon_tracker/pokemon/src/model.dart';
 import 'package:pokemon_tracker/pokemon/widgets/pokemon_asset_view.dart';
-import 'package:pokemon_tracker/pokemon/widgets/pokemon_data_widget.dart';
 
 class PokemonDetailsScreen extends StatelessWidget {
   const PokemonDetailsScreen({super.key});
@@ -19,20 +20,7 @@ class PokemonDetailsScreen extends StatelessWidget {
           return state is PokemonLoaded || state is PokemonLoading || state is PokemonInitial || state is PokemonError;
         },
         builder: (context, state) {
-          if (state is PokemonInitial || state is PokemonLoading) {
-            return Column(
-              children: [
-                PokemonAssetView(frontImage: state.pokemonImageUrl),
-                const Center(child: CircularProgressIndicator()),
-              ],
-            );
-            // } else if (state is PokemonLoading) {
-            //   return _buildLoading();
-          } else if (state is PokemonLoaded) {
-            return PokemonData(pokemon: state.pokemon);
-            // } else if (state is PokemonError) {
-            //   return Container();
-          } else if (state is PokemonError) {
+          if (state is PokemonError) {
             return Column(
               children: [
                 PokemonAssetView(frontImage: state.pokemonImageUrl),
@@ -40,17 +28,135 @@ class PokemonDetailsScreen extends StatelessWidget {
                 Center(child: SelectableText(state.message)),
               ],
             );
-            // } else if (state is PokemonError) {
-            //   return Container();
-          } else {
-            return Container();
           }
+          return SafeArea(
+            child: SingleChildScrollView(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Stack(
+                    alignment: Alignment.topCenter,
+                    children: [
+                      Container(
+                        padding: const EdgeInsets.symmetric(horizontal: 8),
+                        height: 100,
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          mainAxisSize: MainAxisSize.min,
+                          children: [
+                            Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                Text(state is! PokemonLoaded ? '' : state.pokemon.name![0].toUpperCase() + state.pokemon.name!.substring(1), style: Theme.of(context).textTheme.headlineLarge),
+                                Text(state is! PokemonLoaded ? '' : '#${state.pokemon.id}', style: Theme.of(context).textTheme.headlineLarge),
+                                // Text('#${colorSchemeFromImage.background}', style: Theme.of(context).textTheme.headlineLarge),
+                              ],
+                            ),
+                            if (state is PokemonLoaded)
+                              Row(
+                                children: state.pokemon.types!.map((type) {
+                                  return TypeContainer(type: type);
+                                }).toList(),
+                              ),
+                          ],
+                        ),
+                      ),
+                      PokemonAssetView(
+                        backImage: state.pokemonImageUrl,
+                        frontImage: state.pokemonImageUrl,
+                      ),
+                    ],
+                  ),
+                  if (state is PokemonLoaded) StatIndicator(stats: state.pokemon.stats!),
+                ],
+              ),
+            ),
+          );
+          // if (state is PokemonInitial || state is PokemonLoading) {
+          //   return Column(
+          //     children: [
+          //       PokemonAssetView(frontImage: state.pokemonImageUrl),
+          //       const Center(child: CircularProgressIndicator()),
+          //     ],
+          //   );
+          //   // } else if (state is PokemonLoading) {
+          //   //   return _buildLoading();
+          // } else if (state is PokemonLoaded) {
+          //   return PokemonData(pokemon: state.pokemon);
+          //   // } else if (state is PokemonError) {
+          //   //   return Container();
+          // }
         },
       ),
       // floatingActionButton: FloatingActionButton(
       //   onPressed: () => Navigator.of(context)..pop(),
       //   child: const Icon(Icons.arrow_back),
       // ),
+    );
+  }
+}
+
+class TypeContainer extends StatelessWidget {
+  const TypeContainer({
+    required this.type,
+    super.key,
+  });
+
+  final Types type;
+
+  @override
+  Widget build(BuildContext context) {
+    final screenSize = MediaQuery.of(context).size;
+    return SizedBox(
+      width: screenSize.width * 0.12,
+      height: screenSize.height * 0.05,
+      child: CachedNetworkImage(
+        imageUrl: type.type.url,
+        cacheKey: type.type.url,
+        height: screenSize.height * 0.05,
+      ),
+    );
+  }
+}
+
+class StatIndicator extends StatelessWidget {
+  const StatIndicator({
+    required this.stats,
+    super.key,
+  });
+
+  final List<Stats> stats;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.symmetric(horizontal: 8, vertical: 16),
+      width: double.infinity,
+      child: Table(
+        defaultVerticalAlignment: TableCellVerticalAlignment.middle,
+        columnWidths: const {
+          0: FlexColumnWidth(4),
+          1: FlexColumnWidth(),
+          2: FlexColumnWidth(7),
+        },
+        children: [
+          ...stats.map((stat) {
+            return TableRow(
+              children: [
+                TableCell(child: Text(stat.stat.name!.toUpperCase())),
+                TableCell(child: Text(stat.baseStat.toString().trim())),
+                TableCell(
+                  child: LinearProgressIndicator(
+                    value: stat.baseStat! / 100,
+                    semanticsLabel: '${stat.stat.name!.toUpperCase()} indicator',
+                  ),
+                ),
+              ],
+            );
+          }),
+        ],
+      ),
     );
   }
 }
